refactor: improve pkg/rtmp code quality and Go conventions compliance
• Update constants to CamelCase naming (MSG_TYPE_* → MsgType*, etc.)
• Add nil safety checks for unsafe.Pointer usage
• Unify struct field visibility (Timestamp → timestamp)
• Convert private functions to public API (newBasicHeader → NewBasicHeader)
• Add mutex for concurrent access to session state
• Fix HLS FMP4 type conversion issues

+++ pkg/hls/fmp4/writer.go
@@ -541,7 +541,7 @@ func (w *FMP4Writer) writeTFHD(traf *TrackFragment, baseDataOffset uint32) error
 		boxSize += 4
 	}
 	
-	if err := w.writeFullBoxHeader(boxSize, BoxTypeTFHD, 0, flags); err != nil {
+	if err := w.writeFullBoxHeader(boxSize, BoxTypeTFHD, 0, uint32(flags)); err != nil {
 		return err
 	}
 	
@@ -571,7 +571,7 @@ func (w *FMP4Writer) writeTRUN(traf *TrackFragment, dataOffset uint32) error {
 	// 박스 크기 계산
 	boxSize := uint32(8 + 4 + 4 + 4 + len(traf.Samples)*12) // 기본 + 샘플별 12바이트
 	
-	if err := w.writeFullBoxHeader(boxSize, BoxTypeTRUN, 0, flags); err != nil {
+	if err := w.writeFullBoxHeader(boxSize, BoxTypeTRUN, 0, uint32(flags)); err != nil {
 		return err
 	}
